create a new file for class EncryptionConfiguration
