+++ bigquery/google/cloud/bigquery/__init__.py
@@ -73,12 +73,12 @@ from google.cloud.bigquery.routine import Routine
 from google.cloud.bigquery.routine import RoutineArgument
 from google.cloud.bigquery.routine import RoutineReference
 from google.cloud.bigquery.schema import SchemaField
-from google.cloud.bigquery.table import EncryptionConfiguration
 from google.cloud.bigquery.table import Table
 from google.cloud.bigquery.table import TableReference
 from google.cloud.bigquery.table import Row
 from google.cloud.bigquery.table import TimePartitioningType
 from google.cloud.bigquery.table import TimePartitioning
+from google.cloud.bigquery.encryption_configuration import EncryptionConfiguration
 
 __all__ = [
     "__version__",
@@ -94,7 +94,6 @@ __all__ = [
     "DatasetReference",
     "AccessEntry",
     # Tables
-    "EncryptionConfiguration",
     "Table",
     "TableReference",
     "Row",
@@ -136,6 +135,8 @@ __all__ = [
     "StandardSqlDataTypes",
     "SourceFormat",
     "WriteDisposition",
+    # EncryptionConfiguration
+    "EncryptionConfiguration",
 ]
