+++ bigquery/tests/unit/test_job.py
@@ -1229,7 +1229,9 @@ class TestLoadJobConfig(unittest.TestCase, _Base):
         self.assertIsNone(config.destination_encryption_configuration)
 
     def test_destination_encryption_configuration_hit(self):
-        from google.cloud.bigquery.table import EncryptionConfiguration
+        from google.cloud.bigquery.encryption_configuration import (
+            EncryptionConfiguration,
+        )
 
         kms_key_name = "kms-key-name"
         encryption_configuration = EncryptionConfiguration(kms_key_name)
@@ -1242,7 +1244,9 @@ class TestLoadJobConfig(unittest.TestCase, _Base):
         )
 
     def test_destination_encryption_configuration_setter(self):
-        from google.cloud.bigquery.table import EncryptionConfiguration
+        from google.cloud.bigquery.encryption_configuration import (
+            EncryptionConfiguration,
+        )
 
         kms_key_name = "kms-key-name"
         encryption_configuration = EncryptionConfiguration(kms_key_name)
@@ -2439,7 +2443,9 @@ class TestCopyJobConfig(unittest.TestCase, _Base):
         self.assertEqual(config.write_disposition, write_disposition)
 
     def test_to_api_repr_with_encryption(self):
-        from google.cloud.bigquery.table import EncryptionConfiguration
+        from google.cloud.bigquery.encryption_configuration import (
+            EncryptionConfiguration,
+        )
 
         config = self._make_one()
         config.destination_encryption_configuration = EncryptionConfiguration(
@@ -3364,7 +3370,9 @@ class TestQueryJobConfig(unittest.TestCase, _Base):
         self.assertEqual(resource["someNewProperty"], "Woohoo, alpha stuff.")
 
     def test_to_api_repr_with_encryption(self):
-        from google.cloud.bigquery.table import EncryptionConfiguration
+        from google.cloud.bigquery.encryption_configuration import (
+            EncryptionConfiguration,
+        )
 
         config = self._make_one()
         config.destination_encryption_configuration = EncryptionConfiguration(
